Fix clippy lint
