+++ crates/noirc_evaluator/src/ssa_refactor/ir/basic_block.rs
@@ -10,7 +10,7 @@ use super::{
 ///
 /// This means that if one instruction is executed in a basic
 /// block, then all instructions are executed. ie single-entry single-exit.
-#[derive(Debug, Default, PartialEq, Eq, Hash, Clone)]
+#[derive(Debug, PartialEq, Eq, Hash, Clone)]
 pub(crate) struct BasicBlock {
     /// Parameters to the basic block.
     parameters: Vec<ValueId>,
