--- crates/noirc_evaluator/src/ssa_refactor/ssa_builder/ssa_builder.rs
@@ -1,18 +1,0 @@
-use crate::ssa_refactor::ir::{
-    function::{Function, FunctionId},
-    map::AtomicCounter,
-};
-
-/// The global context while building the ssa representation.
-/// Because this may be shared across threads, it is synchronized as necessary.
-#[derive(Default)]
-
-pub(crate) struct SsaBuilder {
-    function_count: AtomicCounter<Function>,
-}
-
-impl SsaBuilder {
-    pub(super) fn next_function(&self) -> FunctionId {
-        self.function_count.next()
-    }
-}
